Correctly changed Utlity to a namespace rather than a class, added new method - positionsIntersect to Utility, overloaded vector corrections, object movement initialized, collision detection added to WorldManager

+++ GameEngine/Vector.cpp
@@ -24,8 +24,8 @@ namespace df {
 	}
 
 	Vector::Vector(int init_x, int init_y) {
-		m_x = init_x;
-		m_y = init_y;
+		m_x = static_cast<float>(init_x);
+		m_y = static_cast<float>(init_y);
 	}
